feat: add MaxL2BlockNumber config param (#648)
It required a way to generate certificate until reach a L2 blockNumber. This is useful, for instance, if there are a fork  in L2 and the we need to verify up to a block to be able to update after that.
The `aggsender`  stop depending on  `StopOnFinishedSendingAllCertificates` when reach this final state, also it shows a log like `"all certs send, no more certificates can be sent`.
```
[AggSender]
MaxL2BlockNumber = 0
StopOnFinishedSendingAllCertificates = false
```
If `MaxL2BlockNumber`  is equal to **0** means disabled. If not this is the maximum block that can be included in a certificate.

Co-authored-by: Rachit Sonthalia <54906134+rachit77@users.noreply.github.com>