fix: handle reorg detection edgecase (#1298)
## 🔄 Changes Summary
- Handle an edgecase when we are not able to track previous block reorgs
and we are on the current block. When polling new block, the blockhash
changes after the last ticker, then we are not able to track the block
hash of previous block and new block is added. Now we are always
comparing the latest block header in reorg detection as well